backend/wChat/requirements.txt
psycopg2-binary==2.9.9
PyJWT==2.8.0
bcrypt==4.0.1
requests==2.26.0
cryptography==36.0.0